requirements.txt
fastapi>=0.115
uvicorn>=0.34
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
scrapling>=0.4
sentence-transformers>=3.0
chromadb>=1.0
rank_bm25>=0.2
